Add more documentation to modular ship assembly and separate EveModularObjectModifer into its own file
