+++ trinity/CMakeLists.txt
@@ -578,6 +578,9 @@ set(_SOURCES
     Eve/SpaceObject/Children/EveCloudEditableVolume.cpp
     Eve/SpaceObject/Children/EveCloudEditableVolume.h
     Eve/SpaceObject/Children/EveCloudEditableVolume_Blue.cpp
+    Eve/SpaceObject/Children/EveModularObjectModifier.cpp
+    Eve/SpaceObject/Children/EveModularObjectModifier.h
+    Eve/SpaceObject/Children/EveModularObjectModifier_Blue.cpp
     Eve/SpaceObject/Children/EveSpaceObjectChild.cpp
     Eve/SpaceObject/Children/EveSpaceObjectChild.h
     Eve/SpaceObject/Children/EveSpaceObjectChild_Blue.cpp
